Created module for mvvmfx-guice. Split up the common code for both examples and the starter into its modules.
