+++ examples/mvvmfx-complex-example/pom.xml
@@ -3,7 +3,7 @@
 	<modelVersion>4.0.0</modelVersion>
 	<groupId>de.saxsys</groupId>
 	<artifactId>mvvmfx-complex</artifactId>
-	<version>0.0.1-SNAPSHOT</version>
+	<version>0.1.1-SNAPSHOT</version>
 	<name>Complex Example</name>
 	<properties>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
@@ -47,13 +47,8 @@
 		<dependency>
 			<groupId>de.saxsys</groupId>
 			<artifactId>mvvmFX</artifactId>
-			<version>0.1-alpha-1-SNAPSHOT</version>
+			<version>0.1.1-SNAPSHOT</version>
 		</dependency>
-        <dependency>
-            <groupId>com.cathive.fx</groupId>
-            <artifactId>fx-guice</artifactId>
-            <version>2.1.3</version>
-        </dependency>
         <dependency>
 			<groupId>org.slf4j</groupId>
 			<artifactId>slf4j-simple</artifactId>
