+++ examples/mvvmfx-guice-starter/pom.xml
@@ -27,6 +27,16 @@
 	</repositories>
 
 	<dependencies>
+        <dependency>
+            <groupId>javax.inject</groupId>
+            <artifactId>javax.inject</artifactId>
+            <version>1</version>
+        </dependency>
+        <dependency>
+            <groupId>com.cathive.fx</groupId>
+            <artifactId>fx-guice</artifactId>
+            <version>2.1.3</version>
+        </dependency>
 		<dependency>
 			<groupId>junit</groupId>
 			<artifactId>junit</artifactId>
@@ -46,7 +56,17 @@
 		<dependency>
 			<groupId>de.saxsys</groupId>
 			<artifactId>mvvmFX</artifactId>
-			<version>0.1-alpha-1-SNAPSHOT</version>
+			<version>0.1.1-SNAPSHOT</version>
+		</dependency>
+        <dependency>
+            <groupId>de.saxsys</groupId>
+            <artifactId>mvvmFX-guice</artifactId>
+            <version>0.1.1-SNAPSHOT</version>
+        </dependency>
+		<dependency>
+			<groupId>de.saxsys</groupId>
+			<artifactId>mvvmfx-complex</artifactId>
+			<version>0.1.1-SNAPSHOT</version>
 		</dependency>
 		<dependency>
 			<groupId>org.slf4j</groupId>
