+++ examples/mvvmfx-guice-starter/src/main/java/de/saxsys/jfx/Starter.java
@@ -1,17 +1,17 @@
 package de.saxsys.jfx;
 
-import java.util.List;
-
+import com.google.inject.Module;
+import de.saxsys.jfx.exampleapplication.view.maincontainer.MainContainerView;
+import de.saxsys.jfx.exampleapplication.viewmodel.maincontainer.MainContainerViewModel;
+import de.saxsys.jfx.mvvm.guice.MvvmfxGuiceApplication;
+import de.saxsys.jfx.mvvm.viewloader.ViewLoader;
+import de.saxsys.jfx.mvvm.viewloader.ViewTuple;
+import javafx.scene.Parent;
 import javafx.scene.Scene;
-import javafx.scene.layout.Pane;
 import javafx.stage.Stage;
 
 import javax.inject.Inject;
-
-import com.google.inject.Module;
-
-import de.saxsys.jfx.mvvm.di.guice.MvvmGuiceApplication;
-import de.saxsys.jfx.mvvm.viewloader.ViewLoader;
+import java.util.List;
 
 /**
  * Entry point of the application.
@@ -19,7 +19,7 @@ import de.saxsys.jfx.mvvm.viewloader.ViewLoader;
  * @author sialcasa
  * 
  */
-public class Starter extends MvvmGuiceApplication {
+public class Starter extends MvvmfxGuiceApplication {
 
 	// Get the MVVM View Loader
 	@Inject
@@ -30,15 +30,18 @@ public class Starter extends MvvmGuiceApplication {
 	}
 
 	@Override
-	public void start(final Stage stage) throws Exception {
-		final Scene scene = new Scene(new Pane(), 300, 300);
-		stage.setScene(scene);
-		stage.show();
+	public void startMvvmfx(final Stage stage) throws Exception {
+        final ViewTuple<MainContainerViewModel> tuple = viewLoader
+                .loadViewTuple(MainContainerView.class);
+        // Locate View for loaded FXML file
+        final Parent view = tuple.getView();
+
+        final Scene scene = new Scene(view);
+        stage.setScene(scene);
+        stage.show();
 	}
 
-	@Override
-	public void initGuiceModules(List<Module> modules) throws Exception {
-
-	}
-
+    @Override
+    public void initGuiceModules(List<Module> modules) throws Exception {
+    }
 }
