--- examples/mvvmfx-guice-starter/src/main/resources/de/saxsys/jfx/exampleapplication/view/maincontainer/MainContainerView.fxml
@@ -1,22 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import java.lang.*?>
-<?import javafx.scene.control.*?>
-<?import javafx.scene.layout.*?>
-<?import javafx.scene.layout.VBox?>
-
-<Pane id="personInfoVbox" prefHeight="149.0" xmlns:fx="http://javafx.com/fxml" fx:controller="de.saxsys.jfx.exampleapplication.view.maincontainer.MainContainerView">
-  <!-- TODO Add Nodes -->
-  <children>
-    <ScrollPane layoutX="0.0" layoutY="0.0" minWidth="470.0" prefHeight="149.0" prefWidth="470.0">
-      <content>
-        <VBox fx:id="personInfoVbox" prefHeight="-1.0" prefWidth="461.0">
-          <children>
-            <!-- Take care for setting the ID to use the injected objects in the code-behind -->
-            <fx:include fx:id="loginView" source="../personlogin/PersonLoginView.fxml" />
-          </children>
-        </VBox>
-      </content>
-    </ScrollPane>
-  </children>
-</Pane>
