--- examples/mvvmfx-guice-starter/src/main/resources/de/saxsys/jfx/exampleapplication/view/personlogin/PersonLoginView.fxml
@@ -1,33 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import java.lang.*?>
-<?import javafx.collections.*?>
-<?import javafx.scene.control.*?>
-<?import javafx.scene.layout.*?>
-<?import javafx.scene.layout.StackPane?>
-<?import javafx.scene.text.*?>
-
-<StackPane prefHeight="133.9998779296875" prefWidth="461.0" xmlns:fx="http://javafx.com/fxml" fx:controller="de.saxsys.jfx.exampleapplication.view.personlogin.PersonLoginView">
-  <!-- TODO Add Nodes -->
-  <children>
-    <VBox fx:id="layoutVbox" alignment="CENTER" minHeight="37.0" prefHeight="155.0" prefWidth="461.0" spacing="5.0">
-      <children>
-        <Label alignment="CENTER" prefWidth="270.0" text="Personen" textAlignment="LEFT">
-          <font>
-            <Font size="30.0" />
-          </font>
-        </Label>
-        <ChoiceBox id="persons" fx:id="personsChoiceBox" prefWidth="277.0">
-          <items>
-            <FXCollections fx:factory="observableArrayList">
-              <String fx:value="Item 1" />
-              <String fx:value="Item 2" />
-              <String fx:value="Item 3" />
-            </FXCollections>
-          </items>
-        </ChoiceBox>
-        <Button mnemonicParsing="false" onAction="#loginButtonPressed" text="Login" />
-      </children>
-    </VBox>
-  </children>
-</StackPane>
