+++ mvvmfx-parent/mvvmfx-guice/pom.xml
@@ -36,6 +36,19 @@
 			<artifactId>javafx</artifactId>
 		</dependency>
 
+        <dependency>
+            <groupId>de.saxsys</groupId>
+            <artifactId>mvvmFX</artifactId>
+            <version>0.1.1-SNAPSHOT</version>
+            <scope>provided</scope>
+        </dependency>
+
+
+        <dependency>
+            <groupId>com.cathive.fx</groupId>
+            <artifactId>fx-guice</artifactId>
+            <version>2.1.3</version>
+        </dependency>
 		
 
 		<!-- Testing Frameworks -->
